Binary operation checks.
