+++ analyzer/src/namespace/mod.rs
@@ -1,4 +1,3 @@
 pub mod events;
-pub mod operations;
 pub mod scopes;
 pub mod types;
