+++ compiler/tests/compile_errors.rs
@@ -88,7 +88,11 @@ use std::fs;
     case("type_constructor_from_variable.fe", "NumericLiteralExpected"),
     case("unary_minus_on_bool.fe", "TypeError"),
     case("unexpected_return.fe", "TypeError"),
-    case("aug_assign_non_numeric.fe", "TypeError")
+    case("aug_assign_non_numeric.fe", "TypeError"),
+    case("binary_operations/add_uints.fe", "TypeError"),
+    case("binary_operations/lshift_bool.fe", "TypeError"),
+    case("binary_operations/lshift_with_int.fe", "TypeError"),
+    case("binary_operations/pow_int.fe", "TypeError")
 )]
 fn test_compile_errors(fixture_file: &str, expected_error: &str) {
     let src = fs::read_to_string(format!("tests/fixtures/compile_errors/{}", fixture_file))
