+++ compiler/tests/fixtures/compile_errors/binary_operations/lshift_bool.fe
@@ -1,0 +1,3 @@
+contract Foo:
+    pub def foo(a: bool, b: i256):
+        a << b
