+++ compiler/tests/fixtures/compile_errors/binary_operations/lshift_with_int.fe
@@ -1,0 +1,3 @@
+contract Foo:
+    pub def foo(a: u256, b: i256):
+        a << b
