+++ compiler/tests/fixtures/features/return_bitwiseshr_i256.fe
@@ -1,3 +1,3 @@
 contract Foo:
-    pub def bar(x: i256, y: i256) -> i256:
+    pub def bar(x: i256, y: u256) -> i256:
         return x >> y
