+++ newsfragments/347.internal.md
@@ -1,0 +1,1 @@
+Refactoring of binary operation type checking.
